requirements.txt
# Agent Framework
uagents>=0.22.1

# Trello Integration
py-trello>=0.19.0

# Slack Integration
slack-sdk>=3.27.1

# Utilities
python-dotenv>=1.0.1
requests>=2.31.0
urllib3>=2.2.1

crewai>=0.11.0
pyyaml>=6.0.1
pydantic>=2.5.2
openai>=1.3.0
pysbd>=0.3.4
matplotlib>=3.8.0
seaborn>=0.13.0
langchain>=0.1.0
langchain-openai>=0.0.2
pytz>=2024.1
certifi>=2024.2.2
aiohttp>=3.9.3
cloudinary>=1.43.0
markdown>=3.6
fpdf2>=2.7.9
